Replace aapt_Images_cpp_Res_png_9patch.diff with aapt_Images_cpp_remove_assertions.diff. This patch now removes some assertions instead of patching the API usage.
For unidentified reasons, the assertions in
checkNinePatchSerialization(Res_png_9patch *,  void *) always fail, which
prevents aapt from outputting any APKs or JARS.
If they are not removed, the fact that they are using the wrong API of
Res_png_9patch would also prevent aapt from being successfully compiled.
AOSP has not changed the API but still builds aapt, which means the
assertions are disabled. Hence it is safe to remove them.